+++ pcbnew/dialogs/dialog_graphic_item_properties_for_Modedit.cpp
@@ -134,7 +134,7 @@ bool DIALOG_MODEDIT_FP_BODY_ITEM_PROPERTIES::TransferDataToWindow()
         m_DefaulThicknessTextUnit,
     };
 
-    for( int ii = 0; ii < DIM( texts_unit ); ii++ )
+    for( size_t ii = 0; ii < DIM( texts_unit ); ii++ )
     {
         texts_unit[ii]->SetLabel( GetAbbreviatedUnitsLabel() );
     }
